requirements.txt
annotated-types==0.6.0
anyio==3.7.1
autopep8==2.0.4
azure-appconfiguration==1.5.0
azure-core==1.29.5
azure-functions==1.17.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
cssselect==1.2.0
cssutils==2.9.0
distro==1.8.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
isodate==0.6.1
lxml==4.9.3
openai==1.3.5
packaging==23.2
platformdirs==4.0.0
premailer==3.10.0
pycodestyle==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
pytoolconfig==1.2.6
requests==2.31.0
rope==1.11.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
yagmail==0.15.293
